requirements.txt
authlib ~= 1.0
django ~= 4.0
python-dotenv ~= 0.19
requests ~= 2.27
pymongo ~= 4.8.0
numpy ~= 1.24.4
pandas ~= 2.0.3
gunicorn ~= 21.2.0
yfinance
MarkupSafe
urllib3
Werkzeug
bs4
pytz
ParameterSource
streamlit
langchain-community
langchain_openai
regex
matplotlib